Store frame offsets relative to the previous frame and shrink to 16bit
Taking advantage of a 2 byte alignment, this allows 128k between frames

+++ avi-file.hpp
@@ -72,7 +72,7 @@ struct Stream
 
     uint32_t curFrame = 0;
     uint32_t curOffset = 0;
-    std::vector<uint32_t> frameOffsets;
+    std::vector<uint16_t> frameOffsets;
 };
 
 enum class AudioFormat
